Rename modules for greater clarity
This renames modules and objects to make the distinction between
Dirichlet distributions and Dirichlet processes clearer, as well
as the difference between Dirichlet families and hierarchical
Dirichlet distributions / processes.
