+++ tests/hierarchical_dirichlet_process/test_dirichlet_process_family.py
