+++ tests/hierarchical_dirichlet_process/test_hierarchical_dirichlet_process.py
