+++ tests/hierarchical_dirichlet_process/test_states.py
